move sort menu and filter menu to their own components
